Introduce Depthmap class (#367)
Motivation: A lot of things belong together in depthmap.py. We need to pass around the same variables over and over: width, height, intrinsic, matrix

Solution:
We create a class Depthmap which takes the variables as instance variables so they can be accessed inside the class. Instead of passing the variables we can pass Depthmap objects.